+++ src/form/field-decoration/types.ts
@@ -1,0 +1,22 @@
+export enum FieldDecorationType {
+    UNDERLINE,
+    FLOATING_LABEL
+}
+
+export interface FieldDecorationBuilderProps {
+    onFieldChange: (value: boolean) => void;
+    onFieldFocus: () => void;
+    onFieldBlur: () => void;
+}
+
+export type FieldDecorationBuilder = (hooks: FieldDecorationBuilderProps) => any;
+
+export interface FieldDecoration {
+    label?: any;
+    error?: any;
+    leading?: any;
+    trailing?: any;
+    disabled?: boolean;
+    decoration?: FieldDecorationType;
+    children: FieldDecorationBuilder;
+}