+++ src/providers/flash-provider/index.stories.tsx
@@ -8,7 +8,7 @@ export default {
     component: FlashProvider
 };
 
-export const usage = (): any => {
+export const normal = (): any => {
     const ExampleApp = (): any => {
         const { flashInfo, flashWarning, flashError, flashSuccess } = useContext(FlashProviderContext);
         const message = "Donec_nec_justo_eget_felis_facilisis_fermentum. Aliquam porttitor mauris sit amet orci. dignissim pellentesque felis.";